api/requirements.txt
click==7.1.2
fastapi==0.58.0
h11==0.9.0
httptools==0.1.1
psycopg2-binary==2.8.5
pydantic==1.5.1
starlette==0.13.4
uvicorn==0.11.5
uvloop==0.14.0
websockets==8.1
